Update GWG membership and add a statement about proxies
Doug Marschke was voted in as a member of the Governance Working Group on
September 12th 2017 but we forgot to update the member list in the repo.

This update also proposes a policy statement that members who will miss a
meeting may specify a proxy to represent them for quorum and voting.

Partial-Bug: 1708708

Change-Id: Ic80b410e4b189aa495977d9056e367aa4f352793